fix: fix bugs in layoutelements (#393)
This PR fixes two bugs:

- fix a type casting issue when subtracting a int array with a float.
This popped up when testing with `unstructured`, and some sources of
element coordinates are of `int` type. This PR adds a new unit test case
for `int` coord type with the grouping function
- fix element class id 0 becomes None bug: this happens when dumping
`LayoutElements` as a list of `LayoutElement`. When an element class id
is 0 the logic on main would treat it as no existing and use `None` as
the type.

+++ unstructured_inference/models/yolox.py
@@ -136,7 +136,7 @@ class UnstructuredYoloXModel(UnstructuredObjectDetectionModel):
         sorted_dets = dets[order]
 
         return LayoutElements(
-            element_coords=sorted_dets[:, :4],
+            element_coords=sorted_dets[:, :4].astype(float),
             element_probs=sorted_dets[:, 4].astype(float),
             element_class_ids=sorted_dets[:, 5].astype(int),
             element_class_id_map=self.layout_classes,
